requirements.txt
# Core
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1

# OpenAI
openai==1.68.0

# Document parsing
PyMuPDF==1.25.3
pandas>=2.2.3
openpyxl==3.1.5

# Vector search (semantic / RAG)
chromadb>=1.0.0

# Gmail API
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.163.0

# Config
python-dotenv==1.0.1

# Frontend
streamlit==1.43.1
